Extract JSON store processing, overall process
* JSON processing can move to specific GitHubJsonStore (theoretically we
could support more stores and more than just GitHub)
* DockerfileGitHubUtil could potentially start migrating to an
interface to provide a dry run mode (but it has far too many
responsibilities at the moment)
* Pull generic algorithm out of Parent into GitHubPullRequestSender
